+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - table cells using wrong date formats
+- `requiredAll` options validation when options are hidden
 
 ## [5.12.0] - 2024-03-14